Move `etw` parser to the `execute` module
